Create Reverse String using Advanced Space Techniques
" " (a single space) is represented by "-" (a single dash) in this description. Commands are in quotation marks for extra comprehension.
"POP--var-" instantiates var as an empty string.
"POP--var-string" puts all of the string except the first character in var without saving the first character anywhere else.
"GOTO-label1" goes to label1 regardless of truth.
"GOTO-label1-[var]" goes to label1 if var is an empty string.