Create Reverse String using Advanced Space Techniques
" " (a single space) is represented by "-" (a single dash) in this description. Commands are in quotation marks for extra understanding.
"POP--var-" instantiates var as an empty string.
"POP--var-string" puts all the characters of the string except the first character in var. It saves the first character in a variable of the empty string.
"GOTO-label1-[var]-" goes to label1 if var is an empty string.
"GOTO-label1" goes to label1 without checking anything, like "GOTO-label1-a-a".